fix: widen the IRQ-off accumulator to 64 bits
The running total of IRQ-off microseconds was uint32, which wraps at ~4295 s of
accumulated stall. That is not reachable in a save-storm, but it is exactly the
kind of thing a 72 h soak reaches — and the failure is silent: the counter keeps
counting and the reported average quietly becomes wrong, in the one run the
metric exists to serve.

Widening is free on this path. Verified in the disassembly that the wrapper
still contains a single bl, to flash_range_erase at 0x20000978 (SRAM): the
64-bit add inlines to adds/adcs, the timer read is a direct load from
0x40054000+0x28, and every literal-pool word resolves into RAM. No libgcc
helper is reached, so the OTA-apply path stays safe. The division moves to
CommandManager, which is ordinary flash-resident code.

+++ src/CommandManager.cpp
@@ -825,7 +825,7 @@ void CommandManager::renderMetrics( ) {
   consolePrintf (pt ? " IRQ-off max: %lu us | media: %lu us\n"
                     : " IRQ-off max: %lu us | avg: %lu us\n",
   (unsigned long)g_flashIrqMaxUs,
-  (unsigned long)(irqOps ? g_flashIrqTotalUs / irqOps : 0));
+  (unsigned long)(irqOps ? (uint32_t)(g_flashIrqTotalUs / irqOps) : 0u));
   consolePrintf (pt ? " IRQ-off erase: %lu | prog: %lu | >1ms: %lu\n"
                     : " IRQ-off erase: %lu | prog: %lu | >1ms: %lu\n",
   (unsigned long)g_flashIrqEraseCount,
